Add WeatherScaleView to display the difference in degrees and refactoring

+++ TheTestWeather/Extensions/Float+Extensions.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Double {
+extension Float {
     func toInt() -> Int {
         Int(self)
     }
